requirements.txt
openai
python-pptx
Pillow
pdf2image
moviepy